requirements.txt
pandas==2.2.3
sqlalchemy==2.0.41
pymysql==1.1.1
fastapi==0.115.12
uvicorn==0.34.3
scikit-learn==1.6.1
faiss-cpu==1.11.0.post1
cryptography==45.0.5
torch==2.7.1+cu118
huggingface-hub==0.35.1
tokenizers==0.22.1
transformers==4.56.2
accelerate==1.10.1
safetensors==0.5.3
bitsandbytes==0.47.0